Move hosts file parsing to urlParser
