+++ js/searchbar/hostsSuggestionsPlugin.js
@@ -1,25 +1,4 @@
-var fs = require('fs');
-var hosts = [];
-
-var HOSTS_FILE = process.platform === 'win32'
-    ? 'C:/Windows/System32/drivers/etc/hosts'
-    : '/etc/hosts';
-
-fs.readFile(HOSTS_FILE, 'utf8', parseHosts);
-
-function parseHosts(err, data) {
-    if (err) {
-        return;
-    }
-
-    data = data.replace(/(\s|#.*|255\.255\.255\.255|broadcasthost)+/g, ' ').split(' ');
-
-    data.forEach(function(host) {
-        if (host.length > 0 && hosts.indexOf(host) === -1) {
-            hosts.push(host);
-        }
-    });
-}
+// hosts are parsed in util/urlParser
 
 function showHostsSuggestions(text, input, event, container) {
 
@@ -47,7 +26,7 @@ function showHostsSuggestions(text, input, event, container) {
 registerSearchbarPlugin("hostsSuggestions", {
     index: 1,
     trigger: function (text) {
-        return (typeof text === 'string' && text.length > 2 );
+        return (hosts.length && typeof text === 'string' && text.length > 2 );
     },
     showResults: showHostsSuggestions,
 });
